trie: widen scores from float to double end-to-end
TrieNode.score, TrieNode.maxChildScore, TrieIterator.minScore and
TrieSearchResult.score change from float to double, and all public
trie APIs that take/return a score do the same. Internal API already
took double and narrowed at the boundary; this drops those narrowing
casts and lets precision flow through Trie_Search's exp/sqrt scoring
and back out to FT.SUGGET WITHSCORES.

Cost: +8 bytes per TrieNode. RDB format unchanged (already 8-byte
SaveDouble/LoadDouble on disk).

All callers updated: spec, query, spell_check, dictionary, fork_gc,
debug_commands. Tests in test_trie.c and test_cpp_trie.cpp updated to
match the wider signatures.

+++ src/debug_commands.c
@@ -452,7 +452,7 @@ DEBUG_COMMAND(DumpTerms) {
 
   rune *rstr = NULL;
   t_len slen = 0;
-  float score = 0;
+  double score = 0;
   int dist = 0;
   size_t termLen;
 
@@ -796,7 +796,7 @@ DEBUG_COMMAND(DumpSuffix) {
     TrieIterator *it = Trie_IterateAll(suffix);
     rune *rstr;
     t_len len;
-    float score;
+    double score;
 
     while (TrieIterator_Next(it, &rstr, &len, NULL, &score, NULL, NULL)) {
       size_t slen;
